Crate libp2p_autonat

Implementation of the AutoNAT protocol.

Structs

Behaviour

NetworkBehaviour for AutoNAT.

Config

Config for the Behaviour.

ProbeId

Unique identifier for a probe.

Enums

Event

Event produced by Behaviour.

InboundFailure

Possible failures occurring in the context of receiving an inbound request and sending a response.

InboundProbeError

Inbound probe failed.

InboundProbeEvent

NatStatus

Assumed NAT status.

OutboundFailure

Possible failures occurring in the context of sending an outbound request and receiving the response.

OutboundProbeError

Outbound probe failed or was aborted.

OutboundProbeEvent

ResponseError