Struct libp2p_autonat::Behaviour

pub struct Behaviour { /* fields omitted */ }

NetworkBehaviour for AutoNAT.

The behaviour frequently runs probes to determine whether the local peer is behind NAT and/ or a firewall, or publicly reachable. In a probe, a dial-back request is sent to a peer that is randomly selected from the list of fixed servers and connected peers. Upon receiving a dial-back request, the remote tries to dial the included addresses. When a first address was successfully dialed, a status Ok will be send back together with the dialed address. If no address can be reached a dial-error is send back. Based on the received response, the sender assumes themselves to be public or private. The status is retried in a frequency of Config::retry_interval or Config::retry_interval, depending on whether enough confidence in the assumed NAT status was reached or not. The confidence increases each time a probe confirms the assumed status, and decreases if a different status is reported. If the confidence is 0, the status is flipped and the Behaviour will report the new status in an OutEvent.

Implementations

impl Behaviour

pub fn new(local_peer_id: PeerId, config: Config) -> Self

pub fn public_address(&self) -> Option<&Multiaddr>

Assumed public address of the local peer. Returns None in case of status NatStatus::Private or NatStatus::Unknown.

pub fn nat_status(&self) -> NatStatus

Assumed NAT status.

pub fn confidence(&self) -> usize

Confidence in the assumed NAT status.

pub fn add_server(&mut self, peer: PeerId, address: Option<Multiaddr>)

Add a peer to the list over servers that may be used for probes. These peers are used for dial-request even if they are currently not connection, in which case a connection will be establish before sending the dial-request.

pub fn remove_server(&mut self, peer: &PeerId)

Remove a peer from the list of servers. See Behaviour::add_server for more info.